Stas's Reduction of Jump

+++ Assets/Scenes/Scripts/ScoreMapping.cs
@@ -14,6 +14,6 @@ public class ScoreMapping : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        GetComponent<Text>().text = PlayerPrefs.GetInt("CurrentZPosition").ToString();
+        GetComponent<Text>().text = ((int)PlayerPrefs.GetFloat("CurrentZPosition")).ToString();
     }
 }
